made useOutsideAlerter hook from function and imported color from theme
