+++ app/src/hooks/useOutsideAlerter.tsx
@@ -1,0 +1,23 @@
+import { useEffect } from 'react'
+
+// eslint-disable-next-line react/display-name
+export const useOutsideAlerter = (setState: any, ref: any) => {
+  useEffect(() => {
+    /**
+     * Alert if clicked on outside of element
+     */
+    function handleClickOutside(event: any) {
+      if (ref.current && !ref.current.contains(event.target)) {
+        setState(false)
+      }
+    }
+
+    // Bind the event listener
+    document.addEventListener('mousedown', handleClickOutside)
+    return () => {
+      // Unbind the event listener on clean up
+      document.removeEventListener('mousedown', handleClickOutside)
+    }
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, [ref])
+}
